`PwBaseWorkChain`: make `overrides` absolute
When trying to use the `PdosWorkChain.get_builder_from_protocol()`
method with `ElectronicType.INSULATOR`, the method will fail because the
choice of `ElectronicType` will set `SYSTEM.occupations` to `'fixed'`
for the `nscf` calculation, which will not pass the validation of the
`PdosWorkChain`.

The reason for this is the fact that the `overrides` of the
`PdosWorkChain` protocol file is overwritten by the choice of parameters
set by the `ElectronicType`. To fix this issue, and also move towards
the principles outlined in #745, we now ensure that the `overrides` have
the final say for the `pw.parameters` and `pw.pseudos` inputs.

This also means that the `starting_magnetization` specified in the
`overrides` also takes precedence over the `initial_magnetic_moments`
inputs argument of the `get_builder_from_protocol()` method.

We update the corresponding test and add tests for the
`PdosWorkChain.get_builder_from_protocol()` method.