`PhBaseWorkChain`: add handler for `ERROR_SCHEDULER_OUT_OF_WALLTIME` (#754)
Certain scheduler plugins can detect an out-of-walltime error in which
case the `ERROR_SCHEDULER_OUT_OF_WALLTIME` exit code will already have
been set on the node when the actual output parser is called. The
`PhParser` is updated to check for this exit code, and after having
parsed as much as possible from the output, the same exit code is kept
by not returning any other more specific exit code.

The `PhBaseWorkChain` adds a new handler for this exit code and will
perform a full restart by setting `recover = False`. It needs to be a
full restart because with an OOW error from the scheduler, the state of
the files on disk are almost certainly corrupt as the scheduler will
have killed the job when it was writing to disk.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>