`PhBaseWorkChain`: add handler for diagonalization errors (#757)
The `PwBaseWorkChain` now has a handler for `ERROR_COMPUTING_CHOLESKY`,
an exit code that was added to `PhCalculation` which is returned if the
calculation failed due to problems with cholesky factorization. The
handler will switch to `cg` diagonalization and restart the calculation.
If the diagonalization already was `cg` it will abort.

The `PhParser` was also fixed to return `ERROR_OUTPUT_STDOUT_INCOMPLETE`
in case the stdout was incomplete and no other more specific errors were
detected.