Fix XML parsing for structures with numbered kinds
In #747 we adjusted the parser for `projwfc.x` to rely on the XML file
instead of the provenance for obtaining the structure. This worked fine
for structures that have chemical symbols as kind names, but once the
kind labels contain numbers the `convert_qe2aiida_structure` failed to
generate the `StructureData` based on the XML content since this case is
not considered.

Here we rename this method to `convert_qe_to_aiida_structure` and make
it more flexible by also adding the case of numbered kind names. This is
simply done by removing any digits from the kind names in the XML to
obtain the corresponding `symbol` of the kind name.