`PwParser`: Add support for QE v7.0 (#774)
Add the XML schema for Quantum ESPRESSO v7.0, along with a corresponding
test for the parsing.

Because the `.xsd` filename in the XML output of QE v7.0 is incorrect,
a check is also added in the `get_schema_filename` function to see if
the user was running QE v7.0. In that case, the correct filename is
returned.

We also use `pytest.mark.parametrize` to condense all the XML schema
tests into one.