Devops: fix the CD workflow
Certain steps accidentally had a `-` in front of the `name` as well as
the `uses` key. This made it such that they were interpreted as two
separate steps and so the first one would lack the required `uses`
attribute.

In addition, the workflow should only run on the `master` branch in case
the tagged commit also occurs on other branches, to prevent running the
workflow more than once.