Devops: fix the nightly build
The nightly build against the `develop` branch of `aiida-core` was
failing at the installation of the Python dependencies when running
against Python 3.7. The reason is that the current `develop` branch has
dropped support for Python 3.7 and the `pip` is then failing to install
a version of `pymatgen` that matches the dependency requirements. The
solution is simply to not run against Python 3.7 anymore.

In addition, as of `aiida-core==2.0` the `develop` branch has been
renamed to `main` so the workflow is updated to install that branch. As
a result, the workflow file is also renamed from `develop` to `nightly`
as the term `develop` is no longer really applicable.

Furthermore, the workflow is updated to match the exact same setup
procedure as the `ci.yml` workflow which has also recently been updated.