`PwCalculation`: allow regex patterns in `detect_important_message` (#771)
The parsing of the stdout of `PwCalculation`s went through the function
`detect_important_message` which compared the output to a list of known
error and warning messages. Here we extend this functionality to also
allow regex patterns.

The direct reason is that for QE v6.8 the error message format in the
case of certain pools not having any k-points changed and was no longer
caught by the fixed pattern. This is now fixed by using a regex pattern
instead.