Parsers: add exception title to `ERROR_UNEXPECTED_PARSER_EXCEPTION`
The exact exception and its stack trace was already added as a log to
the relevant process node, however, most often only the exit message is
shown. Currently this would only show the base message that an
unexpected exception occurred which is not very informative and one
would have to run other commands to find out more. By adding the
exception title in the exit message itself, one has more information
directly.