`PwParser`: Fix trajectory verification for `fixed_coords`
For `relax` and `vc-relax` calculations, the `PwParser` does an extra
check on the `TrajectoryData` output to see if the structure has been
properly converged. However, when comparing the final forces with the
threshold, it does not consider that some site coordinates may have been
fixed with the `fixed_coords` setting.

Here we extract the `fixed_coords` setting from the `PwCalculation`
inputs, and pass it to the `verify_convergence_trajectory` function.
This has been adapted to ignore forces for fixed coordinates by setting
the corresponding forces to zero before comparing them with the force
threshold.