Add the `PhWorkChain` and `PhParallelizeQpointsWorkChain` workchains
These were originally developed in `aiida-quantumespresso-epfl` and are
now made public.

The `PhParallelizeQpointsWorkChain` can be used as a drop-in replacement
of the `PhBaseWorkChain` and will run a separate `PhBaseWorkChain` for
each k-point in the requested mesh. Since these calculations are fully
independent they can be submitted in parallel gaining significant speed
ups in the total run time.

The `PhWorkChain` is a simple wrapper to provides a single interface to
the `PhBaseWorkChain` and `PhParallelizeQpointsWorkChain` toggling the
parallelization with a single input.