`MatDynCalculation`: use validator to check `flfrc` input
In the current approach, the `flfrc` input tag is added to the
`_blocked_keywords` class variable on the fly, since it depends on the
filename of the `force_constants` input. However, mutating class
variables can lead to undesirable behaviour. In this case, the
`_blocked_keywords` list would have multiple items for `flfrc`, which
would cause failures later on because the calculation would fail during
the preparation for submission because it looks like the user it trying
to set one of the blocked keywords.

Here we instead use a validator to make sure that the user hasn't
manually specified the `flfrc` input.

The `get_following_text()` is also factored out, since it's cleaner to
simply add these additions in the `generate_input_file` method.