DevOps: update nightly to show as failed on error
In order to let the final step to always run, which is to post to Slack
in case of a failed step, the previous steps enabled `continue-on-error`.
However, this has the undesirable side-effect that the workflow would be
marked as successful.

Instead, we use `if: always()` which will guarantee that the last step
is run while showing the correct exit status.

In addition, we add a condition for the workflow to be run when a pull
request is opened with changes to the workflow itself.