DevOps: only run Slack notification if tests failed (#819)
The workflow was recently adapted to correct the action being marked as
successful even when the tests failed. This was done by removing the
`continue-on-error` option, which ignores the error. To still have the
Slack option run on error, the `if: always()` condition was added. This
works in the case of an error, but it also made that the step would
always run, even when the tests pass just fine. The solution is to keep
the explicit check on the exit status of the install and tests steps.