DevOps: address deprecation warnings (#820)
Addresses all current deprecation warnings but one. The final remaining
warning comes from the standard library because the `imp` module is
being imported, which is deprecated and will be removed in Python 3.12.

The module is being imported by the `past` library, which in turn is
being imported by the `uncertainties` package that is used by our direct
dependency `pint`. There is an open issue on `uncertainties` for this:

    https://github.com/lebigot/uncertainties/issues/152

however, the package maintainer doesn't seem to be very active.