Add the `XspectraCalculation` and `XspectraParser` plugins (#815)
Add the `XspectraCalculation` and `XspectraParser` plugins, as well as associated
entry points, that allow running the `xspectra.x` post-processing code provided
with distributions of Quantum ESPRESSO.

The XSpectra code functions the same as many other post-processing codes as
it essentially post-processes the results of a `pw.x` calculation. In addition, it
needs the radial part of the core wavefunction of the pseudopotential representing
the excited atom. For this, the `upf2plotcore.sh` script is provided by QE. This script
is too simple to write dedicated parsers for, and it is recommended that these are
run directly, for example through using `aiida-shell`.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>