PwParser: Add support for QE v7.1
Add the XML schema for Quantum ESPRESSO v7.1 `pw.x`, along with a corresponding
test for the parser.

Note that the date of the file found in the `qeschemas` repository:

https://github.com/QEF/qeschemas/tree/master/PW_CPV/previous_schemas

doesn't correspond to the date in the header of the actual XML output.
Since this header is used for determining the correct file name, we
adapted the `.xsd` file name accordingly.