`Q2rCalculation`: Fix copying from `parent_folder`
When providing a `RemoteData` node as a `parent_folder` to the
`Q2rCalculation`, the contents of the `out` folder are now copied to the
`DYN_MAT` folder for the `q2r.x` calculation. This is because for
`RemoteData` input the `_default_parent_output_folder` is used, which is
defined as `out` for the parent `NamelistsCalculation` class, but not
overriden by `Q2rCalculation`.

Here we override the `_default_parent_output_folder` class variable so
the contents of the `DYN_MAT` folder are also copied when a `RemoteData`
input is provided.