`PwCalculation/PhCalculation`: add validation for `pseudos` port
The `pseudos` port was not marked as required, making it possible to
submit a calculation without any pseudos, but this would always fail in
the `prepare_for_submission`.

The port is made required and a validator is added to check that the
pseudos provided match the set of kinds of the `structure` input.