`XspectraBaseWorkChain`: First Working Version of `XspectraBaseWorkChain`
Adds a working version of the `XspectraBaseWorkChain` class. This commit
adds a basic `WorkChain` which uses the `XspectraCalculation` and
`PwCalculation` CalcJob classes to compute the X-ray Absorption Spectrum
(XAS) for the K-edge of a given element. The `WorkChain` requires a
structure as input with at least one element marked by the number 1
(e.g. 'Si1') as well as a corresponding core-hole/GIPAW pseudopotential
pair passed to the `WorkChain` via the overrides dictionary for the
element of interest for XAS simulation.

The `WorkChain` in this form requires the `AiiDA-Shell` plugin in order
to handle the `upf2plotcore.sh` step of the calculation, for which a
code node will need to be configured prior to launching the `WorkChain`
for the first time.