`XspectraParser`: Fix bugs in parsing vector/coord system
Adds information on the coordinate system used in the XSpectra
calculation (either crystallographic or Cartesian). Also addresses a bug
spotted in the stdout of XSpectra itself, where the epsilon/k-vector is
reported by the code during the `read_save_file` subroutine as being in
`[Cartesian frame]` regardless of the coordinate system used by the
parent XSpectra calculation. Because of this, calculations where
`xonly_plot=.true.` cannot meaningfully have metadata for the coordinate
system used by the calculation recorded in the `output_parameters`
dictionary by the parser.

This fix also addresses the fact that the epsilon/k-vector is
essentially reported twice in the event of a calculation restart: once
when the input file is read and a second time when the save file is
read.