Hotfix: Update Previous Commit to `XspectraParser`
Updates commit 9ef620c406943aecca9cf331e292ec9507efeab4

Previous commit stated that when reading the save file, XSpectra prints
the epsilon/k-vector stored in the save file as being in Cartesian
coordinates regardless the coordinate system used in the parent
calculation. This statement was incorrect, as XSpectra always records the
epsilon vector in cartesian coordinates. The parser now correctly
records the `eps_k_vector_coords` entry in the event of re-plot
calculations while skipping the entry for calculation restarts, since
the epsilon/k-vector used is reported with the relevant coordinate system
at the beginning of the stdout.