Dependencies: Add lower requirement `aiida-core>=2.0.4` (#841)
The release `aiida-core==2.0` updated the `local_copy_list` syntax to
allow specifying folders with the promise to copy them recursively.
However, due to a bug, the relative source path was included in the
target, essentially duplicating the nesting. This would cause
calculations that restarted from a `retrieved` folder, such as the
`Q2rCalculation` to fail. The bug was fixed in `aiida-core==2.0.4` so we
set version that as a lower boundary requirement.