Refactor: Replace deprecated `Code` data type
The `Code` data plugin has been deprecated. When using it to compare
types, it should be replaced with `AbstractCode`. For concrete
instances, the `InstalledCode` and `PortableCode` plugins should be used
for the "remote" and "local" variant of `Code`, respectively.

Old documentation is also updated to use `load_code` to load a code
instance instead of the various other old and deprecated methods that
were still being suggested.