`XspectraBaseWorkChain`: Improve Handling of CalcFunctions
Changes the method for sending nodes to `get_powder_spectrum` and
`get_all_spectra` by sending the `spectra` nodes from the output of each
`XspectraCalculation` instead of the pk numbers of the nodes. This has
been made possible due to a recent update to `aiida-core` which allows
for nodes to be packaged into a dynamic namespace via `**kwargs`.

The commit also adds a check prior to `get_powder_spectrum` for the case
where a powder spectrum is requested: if vectors '100' and '010' are
present, but '001' is not, the WorkChain will skip
`get_powder_spectrum` to prevent the CalFunction from encountering an
error. This check is in place to prevent the WorkChain from starting the
CalcFunction and having it fail, instead a warning is added to the process
reports.