`XspectraBaseWorkChain`: Improve `get_builder_from_protocol`
Modifies the `get_builder_from_protocol` method to take various
new inputs:
* upf2plotcore_code: a valid code node for upf2plotcore.sh
* core_hole_pseudos: an optional dictionary which can be used to
set the desired pseudopotentials for the absorbing atom in the structure
as well as all other atoms of the same element type. Functions as an
easier alternative to supplying the desired pseudopotentials via
overrides.

The commit also fixes an oversight in `run_upf2plotcore` where the code
node was not applied correctly to `shell_inputs` and instead relied on a
pre-determined label for the code node. The code now correctly assigns
the upf2plotcore code node supplied in the input.