Add the `options` argument to all `get_builder_from_protocol`
Often a user will need to use the same options for all the jobs that
will be run in a nested workflow, such as specifying the `queue_name` or
the `account`, etc. Currently, the user will have to manually specify
those on the builder, or in the `overrides`, but this means that the
user knows exactly where in the nesting the jobs are.

The `options` argument is added to all `get_builder_from_protocol`
implementations, which takes a dictionary just as one would pass to the
`metadata.options` input for a `CalcJob`. The implementation ensures
that these inputs are recursively set on all the inputs of all the
`CalcJob`s that are part of the nested namespace.