Docs: Create the new scaffolding
The scaffolding is heavily based on that of `aiida-core`. It uses the
`pydata_sphinx_theme` to create large panels on the front page to direct
to any of the major section of the documentation. The division of the
"Get started", "Tutorials", "How-to guides", "Topic guides" and
"Reference guides" is taken from popular open source projects such as
Django and is a battle-tested setup of successful documentation.