Docs: Add tutorials
Add a tutorial on running a `pw.x` calculation through the CLI and the
Python API. Most users will be looking to run such a calculation and it
requires the least setup as most other codes will actually require a
previously completed `pw.x` calculation.

The CLI is presented first as it will just require the setup of a code
and pseudopotential family and will automatically create all other
required inputs on the fly.