Docs: Add scaffolding for topics on calculations and workflows
For now the entries are simply stubs with the spec of the process
automatically formatted through the AiiDA Sphinx extension. This will be
useful for the how-to guides to be able to refer to the complete input
and output specification. We should strive to make the documentation of
the process spec in the implementation as complete and clear as possible
so that the automatically generated HTML version is identical and we
don't have multiple places to maintain the specification documentation.