Fix `ph.x` parsing for single q-point list
The PhParser would fail to properly parse the q-points from the stdout
file when only a single q-point was provided as a list at the end of the
input file. This is because the typical summary at the start of the
stdout is no longer there in this case.

Here we add an extra step in the parsing process that checks if the
q-points have already been parsed. If not, the parser now also check for
the single q-point further down the `stdout` file.