`XspectraBaseWorkChain`: Update process flow and fix docstrings
Fixes some oversights and mistakes in the docstrings for various
functions, as well as updating the main docstring with the overall
process flow. Updates the process flow itself to address mistakes picked
up in testing:
* All instances of `collect_powder` changed to `get_powder_spectrum`.
* Updates `get_all_spectra` to collect spin-up and spin-down sigma as
well as the total sigma for spin-polarised calculations.
* Adds requirements for pytests, such as `dry_run` as an input (along
with statements to return inputs in the event of a test).
* Adds `kpoints`, `kpoints_distance`, and `kpoints_force_parity` as
top-level inputs (in addition to a `validate_kpoints` method in the
process flow and required error code 202) to set the kpoint mesh for
the `XspectraCalculation`s specficially (thus no longer using the same
mesh as the `PwBaseWorkChain`).
* Changed `get_builder_from_protocol` to no longer require
`upf2plotcore_code` as input. The same logic used to screen the main inputs
is used in the method in order to ensure that at least one of
`core_wfc_data` or `upf2plotcore_code` is given as input.
* Changed the logic for keeping track of what `XspectraCalculation`s are
to be processed, which are finished, and which need restarting.
`self.ctx.xspectra_calc_labels` stores the root of the label (e.g.
`xas_0`) as a global reference for future processes, the last-submitted
production calculation for a given eps vector will be stored in `self.ctx`
(thus overwriting the last one each time) at the top level (e.g.
`xas_0_prod`, and the current iteration will be kept only in the label
of each subsequent calculation and their link label (e.g.
`xas_0_prod_iter_1`.