`XspectraBaseWorkChain`: Fix issue with `run_upf2plotcore`
Fixes an oversight in the code for `run_upf2plotcore`.

The code mistakenly used the pseudo node for the absorbing `Kind`
instead of the pseudo for the non-absorbing `Kind` of the same element
type. Now a search is performed to get the absorbing `Kind` from the
structure, identify a `Kind` which has the same element type, then
assign the ground-state pseudo of the matched element type in the
`ShellJob` calculation.