`XspectraParser`: Parse parameters from `CalcJob` inputs (#853)
Updates the method for the `XspectraParser` class to parse the following
parameters from the input of the `CalcJob` class rather than the stdout:

 * The polarisation (epsilon) vectors used in the calculation
 * The coordinate system used for the calculation
   (either `crystallographic` or `cartesian`)
 * Whether the calculation was a re-plot or a full calculation (`xonly_plot`)

The keyword `plot_only` in the `output_parameters` is also changed to
`only_plot_spectrum` in order to clarify its meaning.

The changes were made necessary as, during further testing involving
calculation restarts, it was found that XSpectra stores the polarisation
and k-vector used for the calculation in its save file using a cartesian
basis regardless of the setting used in the calculation itself. When
performing a re-start or re-plot calculation the code reports the
polarisation/k-vector stored in the save file, on a cartesian basis,
which causes confusion for the Parser - particularly in the case of a
restarted calculation as the polarisation vector is then reported twice:
once at the initial read of the input file and then again when the save
file is read (and in potentially different coordinate systems, depending
on calculation settings). It was therefore decided to move the parsing
of these parameters from the raw output to just taking them directly
from the `CalcJob` inputs.