Improve q-point data in `output_parameters`
Remove the `num_q_found` from the `output_parameters` and obtain the
`number_of_qpoints` in the `output_parameters` from the sum of
`num_q_found` in the various outputs. The latter is important in case
the outputs are merged from several independant `ph.x` calculations,
e.g. when running several `ph.x` calculations in parallel for different
q-points.