Tests: Add support for any `Data` to `serialize_builder` (#858)
The `serialize_builder` fixture is updated to add support for the
`RemoteData` class as well as adding a fallback for any other `Data`
implementations.

The fallback simply computes the hash of the node as computed by the
`Data.base.caching._get_hash()` method. Note that we cannot use the
public `get_hash` method as that requires the node to be stored which is
not always the case.

For the `RemoteData` a custom solution was necessary as its hash
includes the absolute filepath of the folder as well as the UUID of the
computer, however, these vary from tests to tests and would cause the
comparison to always fail. Instead, the hash of the repository contents
is used which is the interesting part that we want compare.