Fix bug in `wrap_bare_dict_inputs`
From `aiida-core==2.1.0`, the `valid_type` of an input is extended on
the fly with `None` in case the port is not required. This causes an
issue in the `wrap_bare_dict_inputs` function, which checks if the
`valid_type` is _equal_ to `Dict`. Here we adapt the function to check
if `Dict` is one of the valid types after converting the valid types
into a tuple if necessary.