`BasePwCpInputGenerator`: do not require pseudos in `validate_inputs`
The `validate_inputs` of the `BasePwCpInputGenerator` base `CalcJob`
validates that the `structure` and `pseudos` inputs are commnesurate.
However, wrapping processes may decide to exclude these inputs when
exposing the base class in which case they will not be present when the
validator is called. The validator should not return an error message in
this case as the burden of validation rests on the shoulders of the
wrapping process.

The `port_namespace` is passed into the validator call as the third
argument for this very purposes. It can be used to see if the original
ports are still part of the namespace. The validator is updated to skip
validation if either `pseudos` or `structure` are no longer part of the
namespace. Otherwise, validation goes on as usual.