Add the `XspectraBaseWorkChain` (#854)
This `BaseRestartWorkChain` implementation wraps the `XspectraCalculation`
calculation job plugin to add handlers for out-of-walltime errors. Just
like most other base work chains, it provides a builder generator to
easily prepare a builder based on a certain protocol. In addition to the
standard `fast`, `moderate` and `precise` protocols, the `replot` protocol
is added which can be used to restart from a previously completed run
to just replot the final spectrum.