Add Feature: `XspectraCoreWorkChain`
This commit adds a new `XspectraCoreWorkChain`, including basic tests.

The `CoreWorkChain` performs the generic workflow for computing XANES
spectra using `xspectra.x` and `pw.x`:

* Run `pw.x` calculation on a structure with a core-hole present
* Collect the core-wavefunction of the initial state
* Run an `xspectra.x` calculation for each polarisation vector requested
* Return all generated spectra in a single XyData node output
* (Optionally) compute an XAS spectrum of the power sample of the
input structure if the correct basis vectors are provided

In addition to pre-defined protocol inputs, the WorkChain also uses
pre-defined inputs for different core-hole approximations (or
"treatments", as they are referred to in the WorkChain) and follow a
similar infrastructure to the existing protocols system:
* "full" - full-core-hole, where the core-hole state is approximated
using a simple +1 counter charge (default).
* "half" - half-core-hole, where the core-hole state is approximated
using a +0.5 counter charge.
* "xch_fixed" + "xch_smear" - excited-core-hole, where the core-hole
 state is approximated by occupying the first conduction band with
the excited electron (in both "fixed" and "smearing" occupations,
respectively)
* "none" - no core hole, where no approximation is used. To be used only
if no core-hole pseudopotential is needed.