`PwBandsWorkChain`: Fix `nbnd` for spin-polarised calculations
When running spin-polarised calculations (`nspin != 2`) with the
`PwBandsWorkChain` and providing an `nbands_factor`, the number of bands
should _not_ be doubled, since QE doubles the number of treated k-points
instead.

Here we fix the `PwBandsWorkChain` by removing the `nspin_factor` from
the equations to determine the number of bands for the `bands`
calculation in the work chain.