`get_xspectra_structures`: Add functions for molecules
Adds an optional mode for `get_xspectra_structures` to work with
molecular systems using the Pymatgen library. The mode requires
a Bool object `molecule_input` to be included as a keyword argument
(set to `True`). The method will return:
* A new StructureData object for the molecule centred in a periodic box,
 where the periodic dimensions are equal to the greater of 2x the
extent of the molecule in x, y, and z or the `supercell_min_parameter`
 variable.
* A copy of the centred molecule structure for each symmetrically
non-equivalent atomic site with the non-equivalent site marked as the
absorbing atom (using the given `abs_atom_marker`).
* A Dict object containing various data for the input and centred
molecule structures, as well as all information on the symmetrically
non-equivalent sites in the structure.