Add the `XspectraCoreWorkChain` (#872)
The `XspectraCoreWorkChain` takes a given input structure in which one
atomic site is designated as the site for the absorbing atom and run all
`XspectraCalculation`s as requested. The results are post-processed into
a powder spectrum. This WorkChain makes use of as many features already
implemented in `aiida-quantumespresso` as possible, however, it does
require the `aiida-shell` plugin to run `upf2plotcore.sh` to generate
the `core_wfc_data` input.