Add Feature: `XspectraCrystalWorkChain`
Adds a new WorkChain for `xspectra.x` calculations to compute a complete
K-edge XANES for each element of a given structure using `pw.x` and
`xspectra.x`. The WorkChain uses the following process flow:

* Relax the given input structure using the `PwRelaxWorkChain`
(optional).
* Generate the core wavefunction of each ground-state pseudopotential
for each absorbing element considered (requires AiiDA-Shell plugin).
* Analyse the symmetry of the relaxed structure using `spglib`. Use the
outputs of the analysis to prepare inputs for `XspectraCoreWorkChain`
sub-processes.
* Call the `XspectraCoreWorkChain` for each structure prepared in the
previous step to generate the powder sample XANES for each identified
absorbing atom site in the structure.
* Compile all outputs from the `XspectraCoreWorkChain` sub-processes and
return a single XANES spectrum for each element considered in the
analysis.

Excited-state and ground-state pseudopotentials presently must be provided
using the input namespaces `core_hole_pseudos` and `gipaw_pseudos`
respectively, taking the form of a Dict node with the format:

	{'{element} : {'core_hole' : {upf}, 'gipaw' : {upf}}}

In addition to the final XANES spectra for each element, the powder
spectrum of each absorbing atom calculated can be returned as output to
the head WorkChain node with the optional input
`return_all_powder_spectra` (`False` by default).