Tests: Add tests of overrides for `XspectraCoreWorkChain`
Adds a simple test to check that the builder overrides for the `scf` namespace
for the `XspectraCoreWorkChain` are applied correctly.

Since the `scf.pw.parameters` namespace overrides are applied when the
settings for the chosen core-hole treatment are given, the test also
serves to confirm the intended behaviour that the overrides take precedence
over the core-hole treatment defaults - thus ensuring the overrides have
the final say in setting inputs.