`PhParser`: parse symmetry labels from the `stdout` of `ph.x` (#818)
Add parsing of the symmetry labels to the `PhParser`, which are _only_ printed in the `stdout` after the mode frequencies for each computed q-point.

For each q-point, there is already a `dynamical_matrix_<#qpoint>` key in the `output_parameters` output of the `PhCalculation`. Here we add the `mode_symmetry` and `point_group` outputs, parsed from the `stdout`.

As the symmetry labels are _only_ printed for each computed q-point of a `ph.x` run, this adds an extra challenge when not all q-points are calculated in a single run, for example in case the calculation ran into the walltime and hence is restarted by the `PhBaseWorkChain`. This is dealt with by adding a calculation function called `merge_ph_outputs`, which:

* Merges the `dynamic_matrix_X` keys.
* Combines all the `number_of_irr_representations_for_each_q` lists into one.
* Adds all the `number_of_qpoints`.
* Adds all the `wall_time_seconds`.

This calculation function is called in an additional step in the outline of the `PhBaseWorkChain` called `create_merged_output`. This step also:

* Is skipped in case the `ph.x` is only used for the initialization (`only_initialization` == `True`).
* Is skipped in case the `PhBaseWorkChain` only ran one `PhCalculation`.
* Checks that the number of parsed q-points matches the total number, and returns the `ERROR_MERGING_QPOINTS` exit code if this is not the case.

Some additional logic is added for properly parsing `ph.x` calculations that only run a single q-point, for which no list of q-points is provided at the start of the `stdout`. The parsing and output merging also works as expected in case such a run is restarted in a `PhBaseWorkChain`.

The `wall_time` string output is also removed from the `output_parameters`. This serves little purpose and is convoluted to merge.

Co-authored-by: Marnik Bercx <mbercx@gmail.com>