`PwParser`: Add retrieval and parsing of `CRASH` file (#899)
When `pw.x` crashes, the error is written to stdout but also to the
`CRASH` file. So far the parsers were just parsing errors from the
stdout. As of Quantum ESPRESSO v7.2, the errors are written _only_ to
the `CRASH` file.

Here, the `PwCalculation` is updated to always retrieve the `CRASH` file
and the `PwParser` (and relevant raw parsers) are updated to parse its
contents. The parser tests that tested a particular error case, are
duplicated to test parsing from stdout and the `CRASH` files.