`PwBandsWorkChain`: Move inputs to protocol
Several "default" inputs of the `PwBandsWorkChain` are set inside one of the steps of the work chain, which:

1. Can be confusing for the user, since the expected inputs are not there in the input files.
2. Can be frustrating for the user when a different value is desirable, for a use case that may not immediately be obvious.
3. Means default values are specified _both_ in the work chain logic and protocol, making it more difficult to get a clear overview of the input parameters.

Here we move the specification of the default inputs to the protocol file of the `PwBandsWorkChain` (`bands.yaml`). We also explicitly set the restart inputs to make sure the `bands` calculations restarts from the charge density by default.

Since the default protocol now correctly sets the calculation type to `bands`, the `validate_inputs` validator of the `PwCalculation` will raise a warning because a `parent_folder` has not been initially provided. Hence, we set the `validate_inputs_base` validator for the `pw` port of the `bands` namespace, as is also done for e.g. the `nscf` of the `PdosWorkchain`.