`PwBaseWorkChain`: Remove `automatic_parallelization` input
The automatic parallelization feature has been broken for a while, and
is based on dated benchmarks that most likely are not relevant on newer
machines. Since QE v7.1, an automatic parallelization feature has also been
added to `pw.x`.

Here we remove all the code related to the `automatic_parallelization`
input, and deprecate the related exit codes.