Update `validate_hubbard_parameters`
At some point, the function accepted either a node or a pk for the
`hubbard_file` argument (which used to be called `hubbard_file_pk`) and
so in the case of a pk, the node would be loaded and returned. But that
functionality has been removed some time ago and now it only accepts a
node, so there is no need to load and return it. The only code calling
this utility anyway are the various CLI commands that will anyway always
pass a `SinglefileData` node in since that is provided by the
`options.HUBBARD_FILE` which automatically loads the node from the
identifier passed in the command line arguments by the caller.